move scrollTracking to a hook
